requirements.txt
matplotlib==3.5.1
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.0
seaborn==0.11.2
streamlit==1.4.0
